fix: normalize timestamps to UTC, float schemas, WebSocket proxy, 7-day window
- backfill/_to_dt: convert all ENTSO-E timestamps to UTC before DB merge
  to prevent UniqueViolation when Amsterdam tz != stored UTC primary key
- ingestion/_to_dt: same UTC normalization fix
- schemas: Decimal→float in PricePoint, LoadPoint, ProductionPoint
  so JSON serializes as numbers not strings (fixes empty dashboard charts)
- useWebSocket: use window.location.host so wss:// works in Codespaces
- docker-compose: add --timeout-graceful-shutdown 5 to prevent reload hang
- DashboardPage: show last 7 days instead of 5
- test_load.py: add 3 unit tests for fetch_load with mocks

+++ backend/app/schemas/load.py
@@ -1,12 +1,11 @@
 from pydantic import BaseModel
 from datetime import datetime
-from decimal import Decimal
 
 
 class LoadPoint(BaseModel):
     timestamp: datetime
     bidding_zone: str
-    load_mw: Decimal | None
+    load_mw: float | None
     is_forecast: bool
 
     model_config = {"from_attributes": True}
